moves all changes from mod to separate file
